refactor(routes): extract temporary playback routes
